# Use constant-time head lookups in StateC::{L,R}