@honnibal 2022-01-13
Is the memory usage and copy speed still okay? They're less important but they do become relevant for the beam.