@danieldk 2022-01-13
> Is the memory usage and copy speed still okay? They're less important but they do become relevant for the beam.

I'll do a heap profile to look at memory usage. Any ideas how to profile the copy speed?