@honnibal 2022-01-13
I think you can just use the `StateC.copy()` method -- that's what's used in the beam code