@honnibal 2022-01-13
Btw can you also do a quick parser benchmark just to make sure we don't have some unexpected thing (like I don't know, we create lots of state objects because we expect that to be really cheap, and now there's some overhead or something)

I expect this will all be fine, but it's probably good to check.