@danieldk 2022-01-13
Best-of-three runs using `en_core_web_lg`:

Before this PR: 22092
With this PR: 23345

So, even seems to be a bit faster on small docs.