@danieldk 2022-01-13
Looked again at the profile and `StatC::L`/`R` did indeed take up quite a bit of parsing time before this PR, even with small docs (10 sentences per doc).

<img width="940" alt="Screen Shot 2022-01-13 at 14 06 08" src="https://user-images.githubusercontent.com/49398/149335677-a9bd230d-30fc-4fe1-a828-c2ee81e36390.png">