@ofirnk 2022-01-16
@danieldk I'm running it on the [very messy] colab I've shared earlier. I'll conclude and share graphs a bit later!