@ofirnk 2022-01-16
Ok, so analysis about the amazing progress so far:

1. Before your work, the ratio between v3 and v2 was increasing over document length, reaching 5-6x performance degradation.
2. After first PR - Performance ratio *decreased* to 3-4x :clap: 
3. Including *this* PR - performance is with **constant** degradation of ~30% :1st_place_medal: 

So to summarize - I no longer can see quadratic performance - AMAZING work! 

![image](https://user-images.githubusercontent.com/8449713/149657269-3b51248b-a655-4666-85ef-3abda364162e.png)

Explanation:
1. Yellow squares is the ratio of v3.2 vs v2 [goes from ~1 towards 5-6 (and probably beyond) ]
2. Red line is analysis time in v2, and reddis-triangles is the analysis time with this PR :clap: 
3. Blue triangles is the ratio of this PR versus v2 - which proves we're back to LINEAR analysis time