@danieldk 2022-01-16
Thanks a lot for the analysis 👍🏻, this is really useful!