@ofirnk 2022-01-17
> Thanks a lot for the awesome analysis 👍🏻, this is really useful!

With pleasure, the least I can do for this great open source project 🥇

BTW, this can serve as the regression test for future releases.
Speaking of which - is there an ETA for next release with this [rather critical] fix?