@svlandeg 2022-01-20
> With pleasure, the least I can do for this great open source project 🥇

🙏 

> BTW, this can serve as the regression test for future releases. Speaking of which - is there an ETA for next release with this [rather critical] fix?

This fix will be released as part of spaCy 3.3.0. We don't have a scheduled release date for that yet, but it likely won't take months ;-) Keep an eye out on the release page!